Consolidate AbstractSchemaParser and AbstractDataParser into a single class
